backend/requirements.txt
openai-whisper==20231117
moviepy==1.0.3
ffmpeg-python
pydub==0.25.1
spacy==3.7.2
numpy==1.26.4
matplotlib
textblob==0.18.0
sentence-transformers==2.6.1
torch==2.2.2
scikit-learn==1.4.2
xgboost==2.0.3
joblib==1.4.0
ollama==0.1.7
pandas==2.2.2
flask==3.1.1
flask-cors==6.0.0
tokenizers==0.19.1
